requirements.txt
pyrogram
tgcrypto
python-dotenv
motor
httpx
aiohttp
Babel # (still optional)
pymongo[srv]
